refactor(webpack): move dynamic config files
