+++ build-utils/templates/index.ejs
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <% include ../partials/head %>
+  </head>
+  <body>
+    <div id="main" class="flex-container wrapper">
+      <section class="flex-container--column u-center-text">
+        <h1>Vincent Aceto</h1>
+        <p>
+          Hi! My site is under construction, but please don't hesitate to
+          connect.
+        </p>
+        <ul class="social social--inline u-center-text">
+          <li class="social__item">
+            <svg>
+              <use xlink:href="brands.svg#github"></use>
+            </svg>
+            <a href="https://github.com/vinnyA3" target="_blank"> Github</a>
+          </li>
+          <li class="social__item">
+            <svg>
+              <use xlink:href="brands.svg#codepen"></use>
+            </svg>
+            <a href="https://codepen.io/vinnyA3" target="_blank">Codepen</a>
+          </li>
+          <li class="social__item">
+            <svg>
+              <use xlink:href="brands.svg#linkedin"></use>
+            </svg>
+            <a href="https://linkedin.com/in/vinaceto" target="_blank"
+              >Linkedin</a
+            >
+          </li>
+        </ul>
+      </section>
+    </div>
+  </body>
+</html>
